snapshot_disk: Add snapshot_delete cases without '--metadata' option

+++ libvirt/tests/src/virsh_cmd/snapshot/virsh_snapshot_disk.py
@@ -2,10 +2,10 @@ import os
 import logging
 import re
 import tempfile
-
 from autotest.client.shared import error
 from virttest import virsh, qemu_storage, data_dir
 from virttest.libvirt_xml import vm_xml
+from provider import libvirt_version
 
 
 def run(test, params, env):
@@ -23,6 +23,7 @@ def run(test, params, env):
     vm_name = params.get("main_vm", "virt-tests-vm1")
     vm = env.get_vm(vm_name)
     image_format = params.get("snapshot_image_format", "qcow2")
+    snapshot_del_test = "yes" == params.get("snapshot_del_test", "no")
     status_error = ("yes" == params.get("status_error", "no"))
     snapshot_from_xml = ("yes" == params.get("snapshot_from_xml", "no"))
     snapshot_current = ("yes" == params.get("snapshot_current", "no"))
@@ -45,8 +46,9 @@ def run(test, params, env):
     image = qemu_storage.QemuImg(params, tmp_dir, "snapshot_test")
     img_path, _ = image.create(params)
     # Do the attach action.
+    extra = "--persistent --subdriver %s" % image_format
     result = virsh.attach_disk(vm_name, source=img_path, target="vdf",
-                               extra="--persistent --subdriver %s" % image_format)
+                               extra=extra, debug=True)
     if result.exit_status:
         raise error.TestNAError("Failed to attach disk %s to VM."
                                 "Detail: %s." % (img_path, result.stderr))
@@ -54,6 +56,7 @@ def run(test, params, env):
     # Init snapshot_name
     snapshot_name = None
     snapshot_external_disk = []
+    del_status = None
     try:
         # Create snapshot.
         if snapshot_from_xml:
@@ -76,8 +79,8 @@ def run(test, params, env):
                 lines.append("<disk name='%s' snapshot='%s'>\n" %
                              (disk['source'], snapshot_disk))
                 if snapshot_disk == "external":
-                    disk_external = os.path.join(tmp_dir,
-                                                 "%s.snap" % os.path.basename(disk['source']))
+                    snap_path = "%s.snap" % os.path.basename(disk['source'])
+                    disk_external = os.path.join(tmp_dir, snap_path)
                     snapshot_external_disk.append(disk_external)
                     lines.append("<source file='%s'/>\n" % disk_external)
                 lines.append("</disk>\n")
@@ -88,14 +91,28 @@ def run(test, params, env):
             snapshot_xml_file = open(snapshot_xml_path, "w")
             snapshot_xml_file.writelines(lines)
             snapshot_xml_file.close()
+            logging.debug("The xml content for snapshot create is:")
+            with open(snapshot_xml_path, 'r') as fin:
+                logging.debug(fin.read())
             snapshot_result = virsh.snapshot_create(
-                vm_name, ("--xmlfile %s" % snapshot_xml_path))
+                vm_name, ("--xmlfile %s" % snapshot_xml_path), debug=True)
+            out_err = snapshot_result.stderr.strip()
             if snapshot_result.exit_status:
                 if status_error:
                     return
                 else:
+                    if libvirt_version.version_compare(1, 2, 5):
+                        # As commit d2e668e in 1.2.5, internal active snapshot
+                        # without memory state is rejected. Handle it as SKIP
+                        # for now. This could be supportted in future by bug:
+                        # https://bugzilla.redhat.com/show_bug.cgi?id=1103063
+                        if re.search("internal snapshot of a running VM" +
+                                     " must include the memory state",
+                                     out_err):
+                            raise error.TestNAError("Check Bug #1083345, %s" %
+                                                    out_err)
                     raise error.TestFail("Failed to create snapshot. Error:%s."
-                                         % snapshot_result.stderr.strip())
+                                         % out_err)
         else:
             options = ""
             snapshot_result = virsh.snapshot_create(vm_name, options)
@@ -117,6 +134,9 @@ def run(test, params, env):
                 snapshot_xml_file = open(snapshot_xml_path, "w")
                 snapshot_xml_file.writelines(lines)
                 snapshot_xml_file.close()
+                logging.debug("The xml content for snapshot create is:")
+                with open(snapshot_xml_path, 'r') as fin:
+                    logging.debug(fin.read())
                 options += "--redefine %s --current" % snapshot_xml_path
                 if snapshot_result.exit_status:
                     raise error.TestFail("Failed to create snapshot --current."
@@ -124,8 +144,9 @@ def run(test, params, env):
                                          snapshot_result.stderr.strip())
 
         if status_error:
-            raise error.TestFail("Success to create snapshot in negative case\n"
-                                 "Detail: %s" % snapshot_result)
+            if not snapshot_del_test:
+                raise error.TestFail("Success to create snapshot in negative"
+                                     " case\nDetail: %s" % snapshot_result)
 
         # Touch a file in VM.
         if vm.is_dead():
@@ -138,23 +159,40 @@ def run(test, params, env):
         tmp_file_path = tmp_file.name
         tmp_file.close()
 
-        status, output = session.cmd_status_output("touch %s" % tmp_file_path)
+        echo_cmd = "echo SNAPSHOT_DISK_TEST >> %s" % tmp_file_path
+        status, output = session.cmd_status_output(echo_cmd)
+        logging.debug("The echo output in domain is: '%s'", output)
         if status:
-            raise error.TestFail("Touch file in vm failed. %s" % output)
+            raise error.TestFail("'%s' run failed with '%s'" %
+                                 (tmp_file_path, output))
+        status, output = session.cmd_status_output("cat %s" % tmp_file_path)
+        logging.debug("File created with content: '%s'", output)
 
         session.close()
 
         # Destroy vm for snapshot revert.
-        virsh.destroy(vm_name)
+        if not libvirt_version.version_compare(1, 2, 3):
+            virsh.destroy(vm_name)
         # Revert snapshot.
         revert_options = ""
         if snapshot_revert_paused:
             revert_options += " --paused"
         revert_result = virsh.snapshot_revert(vm_name, snapshot_name,
-                                              revert_options)
+                                              revert_options,
+                                              debug=True)
         if revert_result.exit_status:
-            raise error.TestFail(
-                "Revert snapshot failed. %s" % revert_result.stderr.strip())
+            # As commit d410e6f for libvirt 1.2.3, attempts to revert external
+            # snapshots will FAIL with an error "revert to external snapshot
+            # not supported yet". Thus, let's check for that and handle as a
+            # SKIP for now. Check bug:
+            # https://bugzilla.redhat.com/show_bug.cgi?id=1071264
+            if libvirt_version.version_compare(1, 2, 3):
+                if re.search("revert to external snapshot not supported yet",
+                             revert_result.stderr):
+                    raise error.TestNAError(revert_result.stderr.strip())
+            else:
+                raise error.TestFail("Revert snapshot failed. %s" %
+                                     revert_result.stderr.strip())
 
         if vm.is_dead():
             raise error.TestFail("Revert snapshot failed.")
@@ -164,21 +202,40 @@ def run(test, params, env):
                 vm.resume()
             else:
                 raise error.TestFail("Revert command successed, but VM is not "
-                                     "paused after reverting with --paused option.")
+                                     "paused after reverting with --paused"
+                                     "  option.")
         # login vm.
         session = vm.wait_for_login()
         # Check the result of revert.
         status, output = session.cmd_status_output("cat %s" % tmp_file_path)
+        logging.debug("After revert cat file output='%s'", output)
         if not status:
             raise error.TestFail("Tmp file exists, revert failed.")
 
         # Close the session.
         session.close()
 
+        # Test delete snapshot without "--metadata", delete external disk
+        # snapshot will fail for now.
+        # Only do this when snapshot creat succeed which filtered in cfg file.
+        if snapshot_del_test:
+            if snapshot_name:
+                del_result = virsh.snapshot_delete(vm_name, snapshot_name,
+                                                   debug=True,
+                                                   ignore_status=True)
+                del_status = del_result.exit_status
+                if del_status:
+                    if not status_error:
+                        raise error.TestFail("Failed to delete snapshot.")
+                else:
+                    if status_error:
+                        err_msg = "Snapshot delete succeed but expect fail."
+                        raise error.TestFail(err_msg)
+
     finally:
         virsh.detach_disk(vm_name, target="vdf", extra="--persistent")
         image.remove()
-        if snapshot_name:
+        if del_status and snapshot_name:
             virsh.snapshot_delete(vm_name, snapshot_name, "--metadata")
         for disk in snapshot_external_disk:
             if os.path.exists(disk):
